genai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.0",
    "fastapi>=0.120.2",
    "google-generativeai>=0.8.5",
    "ipywidgets>=8.1.7",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.38.0",
]
